requirements.txt
konlpy==0.6.0
streamlit==1.5.0
pyyaml==6.0
